Modify SoftDevice switch to add versions easily
Modified the configuration switch to build applications with SoftDevices other
than S130 (default). Previously, this was achieved by including in the module
a yotta config that defined `YOTTA_CFG_NORDIC_SOFTDEVICE_S110`. However, this
solution was slightly confusing and not scalable to include support for other
SoftDevice versions in the future. Therefore, the definition was replaced by
`YOTTA_CFG_NORDIC_SOFTDEVICE`, which is set to the desired SoftDevice version
string. Currently supported options are S110 and S130. If no definition is
found then the application is built using the default i.e. S130.