Merge targets by compiler to reduce duplication
The code in the old nordic-nrf51822-16k-armcc and nordic-nrf51822-32k-armcc
(similarly for GCC targets) is very similar but for the linker scripts and a
few macros in the toolchain.cmake file. Therefore, the idea is to combine
these targets and produce nordic-nrf51822-gcc and nordic-nrf51822-armcc.
Furthermore, an option YOTTA_CFG_IMAGE_FOTA is added to automatically build
a fota-able image. Finally, the heap warning can now be controlled by the user
by manually setting YOTTA_CFG_IMAGE_HEAP_WARNING_THRESHOLD.

The directory structure of the targets also changes slightly. A new directory
'licenses' is added to include all the licenses. Also, the directory
'bootloader' contains the bootloader hex file that is usable with both S110 and
S130. Also, the switch-softdevice.md file has been removed and will be replaced
by options.md in the future.