Modify toolchain.cmake that wasnt generating hex
Modify the toolchain.cmake file in both targets to use add_custom_command as
POST_BUILD step rather than associated to a target. Otherwise, the hex files
for -combined.hex and -combined-fota.hex would not be generated. The only
way to fix this using a dependency within add_custom_command would require
the addition of another target, that would not be built if not specified
through yotta.